trend_analytics/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
hdbscan==0.8.36
python-dotenv==1.0.0
pydantic==2.5.0
numpy==1.24.3
pandas==2.1.4
sqlalchemy==2.0.23
sentence-transformers==5.3.0
huggingface-hub==1.7.0
transformers==5.1.0
scikit-learn==1.2.2


